refactor: add exception case in addUser
