+++ study/src/main/java/LDHD/study/common/response/ErrorCode.java
@@ -13,7 +13,8 @@ public enum ErrorCode {
     VALIDATION_FAILED(HttpStatus.BAD_REQUEST, "COMMON400", "유효하지 않는 값입니다."),
 
     // USER
-    USER_NOT_FOUND(HttpStatus.NOT_FOUND, "USER400", "존재하지 않는 회원입니다.");
+    USER_NOT_FOUND(HttpStatus.NOT_FOUND, "USER400", "존재하지 않는 회원입니다."),
+    ALREADY_EXISTS(HttpStatus.BAD_REQUEST, "USER401", "이미 존재하는 회원입니다");
 
     private final HttpStatus status;
     private final String code;
